requirements.txt
boto3
botocore
beautifulsoup4
openai
python-dotenv
